+++ pulsar/internal/lookup_service_test.go
@@ -21,7 +21,7 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/apache/pulsar-client-go/pkg/pb"
+	"github.com/apache/pulsar-client-go/pulsar/internal/pb"
 	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/assert"
 )
